Componentizando pagina
